requirements.txt
anyio==4.4.0
asgiref==3.8.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.0.7
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-decouple==3.8
python-telegram-bot==21.4
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
